requirements.txt
jieba
requests
bs4
nltk
bert-serving-server
bert-serving-client